# Conflicts:
#	.github/workflows/pre-release.yml
#	.github/workflows/release.yml
#	.github/workflows/tests-minimal.yml
#	.github/workflows/tests.yml
#	.github/workflows/unit-tests.yml
#	.gitignore
#	build-tools/build.json
#	build-tools/composer.json
#	build-tools/composer.lock
#	changelogs/minor.md
#	changelogs/patch.md
#	cp-styles/app/styles/components/_fieldsets.scss
#	cp-styles/app/styles/components/_main-nav.scss
#	cp-styles/build/vendor.js
#	phpcs.ruleset.xml
#	system/ee/ExpressionEngine/Addons/channel/mod.channel.php
#	system/ee/ExpressionEngine/Addons/file/ft.file.php
#	system/ee/ExpressionEngine/Addons/member/mod.member.php
#	system/ee/ExpressionEngine/Addons/member/mod.member_settings.php
#	system/ee/ExpressionEngine/Addons/relationship/ft.relationship.php
#	system/ee/ExpressionEngine/Addons/relationship/libraries/EntryList.php
#	system/ee/ExpressionEngine/Addons/search/mod.search.php
#	system/ee/ExpressionEngine/Cli/Commands/CommandUpdate.php
#	system/ee/ExpressionEngine/Cli/Commands/Upgrade/UpgradeMap.php
#	system/ee/ExpressionEngine/Controller/Categories/Categories.php
#	system/ee/ExpressionEngine/Controller/Channels/Channels.php
#	system/ee/ExpressionEngine/Controller/Jumps/Comments.php
#	system/ee/ExpressionEngine/Controller/Members/Members.php
#	system/ee/ExpressionEngine/Controller/Members/Profile/Pro.php
#	system/ee/ExpressionEngine/Controller/Members/Roles/Roles.php
#	system/ee/ExpressionEngine/Controller/Utilities/Communicate.php
#	system/ee/ExpressionEngine/Controller/Utilities/MemberImport.php
#	system/ee/ExpressionEngine/Library/CP/EntryManager/Columns/EntryDate.php
#	system/ee/ExpressionEngine/Library/CP/EntryManager/Columns/ExpirationDate.php
#	system/ee/ExpressionEngine/Library/CP/EntryManager/Columns/Sticky.php
#	system/ee/ExpressionEngine/Library/CP/URL.php
#	system/ee/ExpressionEngine/Library/Data/SerializableEntity.php
#	system/ee/ExpressionEngine/Model/Member/Gateway/MemberGateway.php
#	system/ee/ExpressionEngine/Model/Member/Member.php
#	system/ee/ExpressionEngine/Model/Role/Role.php
#	system/ee/ExpressionEngine/Model/Template/TemplateGroup.php
#	system/ee/ExpressionEngine/Service/Channel/ChannelEntry.php
#	system/ee/ExpressionEngine/Service/Generator/AddonGenerator.php
#	system/ee/ExpressionEngine/Service/LivePreview/LivePreview.php
#	system/ee/ExpressionEngine/Service/Member/Member.php
#	system/ee/ExpressionEngine/Service/Model/Model.php
#	system/ee/ExpressionEngine/Service/Validation/Rule/AlphaDashPeriodEmoji.php
#	system/ee/ExpressionEngine/Service/Validation/Rule/PasswordMatchesSecurityPolicy.php
#	system/ee/ExpressionEngine/Tests/bootstrap.php
#	system/ee/ExpressionEngine/View/_shared/form/field.php
#	system/ee/ExpressionEngine/View/publish/comments/partials/author_information.php
#	system/ee/ExpressionEngine/View/publish/partials/edit_list_table.php
#	system/ee/installer/controllers/wizard.php
#	system/ee/installer/schema/mysql_schema.php
#	system/ee/installer/updates/ud_4_00_05.php
#	system/ee/installer/updates/ud_6_01_00_rc_2.php
#	system/ee/installer/updates/ud_6_02_00.php
#	system/ee/language/english/fieldtypes_lang.php
#	system/ee/language/english/settings_lang.php
#	system/ee/legacy/controllers/cp/login.php
#	system/ee/legacy/libraries/Auth.php
#	system/ee/legacy/libraries/Core.php
#	system/ee/legacy/libraries/Form_validation.php
#	system/ee/legacy/libraries/Stylesheet.php
#	system/ee/legacy/libraries/Template.php
#	system/ee/legacy/libraries/Validate.php
#	tests/cypress/cypress.json
#	tests/cypress/cypress/integration/channel/categories.ee6.js
#	tests/cypress/cypress/integration/content/default_theme.ee6.js
#	tests/cypress/cypress/integration/cp/login.ee6.js
#	tests/cypress/cypress/integration/field/modifiers.ee6.js
#	tests/cypress/cypress/integration/heavyload/many_members.ee6.js
#	tests/cypress/cypress/integration/members/register.ee6.js
#	tests/cypress/cypress/integration/publish/edit.ee6.js
#	tests/cypress/cypress/integration/publish/relationship.ee6.js
#	tests/cypress/cypress/integration/spam/spam_trap.ee6.js
#	tests/cypress/cypress/integration/z_installer/updater.ee6.js
#	tests/cypress/cypress/plugins/index.js
#	tests/cypress/cypress/support/commands.js
#	tests/cypress/members.cypress.json
#	tests/cypress/misc.cypress.json
#	tests/cypress/package-lock.json
#	tests/cypress/package.json
#	tests/cypress/support/config/config.php
#	tests/cypress/support/fixtures/initDb.php
#	tests/cypress/support/fixtures/member.php
#	tests/cypress/support/sql/database_6.0.0.sql
#	tests/cypress/support/sql/more-roles.sql
#	tests/cypress/support/templates/default_site/cats.group/archive-nested.html
#	tests/cypress/support/templates/default_site/cats.group/manual-heading-2.html
#	themes/ee/asset/javascript/src/common.js
#	themes/ee/asset/javascript/src/cp/form_validation.js
#	themes/ee/asset/javascript/src/cp/grid.js
#	themes/ee/asset/javascript/src/cp/passwords.js
#	themes/ee/asset/javascript/src/fields/rte/redactor/redactor.min.js
#	themes/ee/asset/javascript/src/fields/rte/redactor/redactor.min.js.map
#	themes/ee/cp/css/common.min.css
#	themes/ee/cp/js/build/components/relationship.js
#	themes/ee/cp/js/src/components/relationship.jsx
#	themes/ee/debug/css/eecms-debug.min.css

+++ system/ee/installer/updates/ud_6_00_00_b_1.php
@@ -4,7 +4,7 @@
  * ExpressionEngine (https://expressionengine.com)
  *
  * @link      https://expressionengine.com/
- * @copyright Copyright (c) 2003-2021, Packet Tide, LLC (https://www.packettide.com)
+ * @copyright Copyright (c) 2003-2023, Packet Tide, LLC (https://www.packettide.com)
  * @license   https://expressionengine.com/license Licensed under Apache License, Version 2.0
  */
 
@@ -206,7 +206,9 @@ class Updater
         ee('Model')->get('Config')->filter('key', 'IN', ['enable_avatars', 'allow_avatar_uploads'])->delete();
 
         // Remove avatar member preference
-        ee()->dbforge->drop_column('members', 'display_avatars');
+        if (ee()->db->field_exists('display_avatars', 'members')) {
+            ee()->dbforge->drop_column('members', 'display_avatars');
+        }
     }
 
     private function removeJqueryAddon()
